Significantly Cleanup Code
