+++ src/frontend/src/pages/GanttPage/GanttChart.tsx
@@ -1,42 +1,39 @@
 import { Box } from '@mui/material';
-import { ProjectPreview, Team, WorkPackage } from 'shared';
-import { filterGanttProjects, GanttChange, GanttFilters, RequestEventChange } from '../../utils/gantt.utils';
+import { Dispatch } from 'react';
+import { Project, Team, WbsElement, WorkPackage } from 'shared';
+import { filterGanttProjects, GanttFilters } from '../../utils/gantt.utils';
 import GanttChartTeamSection from './GanttChartTeamSection';
 
 interface GanttChartProps {
   startDate: Date;
   endDate: Date;
   teamsList: Team[];
-  saveChanges: (eventChanges: GanttChange[]) => void;
-  showWorkPackagesMap: Map<string, boolean>;
-  setShowWorkPackagesMap: React.Dispatch<React.SetStateAction<Map<string, boolean>>>;
-  highlightedChange?: RequestEventChange;
-  addProject: (project: ProjectPreview, team: Team) => void;
-  getNewProjectNumber: (carNumber: number) => number;
-  addWorkPackage: (workPackage: WorkPackage) => void;
-  getNewWorkPackageNumber: (projectId: string) => number;
   defaultGanttFilters: GanttFilters;
   searchText: string;
-  clearEdits: boolean;
-  setClearEdits: (clearEdits: boolean) => void;
+  addNewWorkPackage: (workPackage: WorkPackage) => void;
+  addNewProject: (project: Project) => void;
+  getNewProjectNumber: (carNumber: number) => number;
+  allWbsElements: WbsElement[];
+  showWorkPackagesMap: Map<string, boolean>;
+  setShowWorkPackagesMap: Dispatch<React.SetStateAction<Map<string, boolean>>>;
+  removeAddedProjects: (projects: Project[]) => void;
+  removeAddedWorkPackages: (workPackages: WorkPackage[]) => void;
 }
 
 const GanttChart = ({
   startDate,
   endDate,
   teamsList,
-  saveChanges,
-  showWorkPackagesMap,
-  setShowWorkPackagesMap,
-  highlightedChange,
-  addProject,
-  addWorkPackage,
-  getNewProjectNumber,
-  getNewWorkPackageNumber,
   defaultGanttFilters,
   searchText,
-  clearEdits,
-  setClearEdits
+  addNewWorkPackage,
+  addNewProject,
+  getNewProjectNumber,
+  allWbsElements,
+  showWorkPackagesMap,
+  setShowWorkPackagesMap,
+  removeAddedProjects,
+  removeAddedWorkPackages
 }: GanttChartProps) => {
   return (
     <Box>
@@ -45,26 +42,20 @@ const GanttChart = ({
 
         const filteredProjects = filterGanttProjects(projects, defaultGanttFilters, searchText, team);
 
-        const addProjectHandler = (project: ProjectPreview) => {
-          addProject(project, team);
-        };
-
         return filteredProjects ? (
           <GanttChartTeamSection
             startDate={startDate}
             endDate={endDate}
-            saveChanges={saveChanges}
-            showWorkPackagesMap={showWorkPackagesMap}
-            setShowWorkPackagesMap={setShowWorkPackagesMap}
             team={team}
             projects={filteredProjects}
-            highlightedChange={highlightedChange}
-            addProject={addProjectHandler}
-            addWorkPackage={addWorkPackage}
+            addNewWorkPackage={addNewWorkPackage}
+            addNewProject={addNewProject}
             getNewProjectNumber={getNewProjectNumber}
-            getNewWorkPackageNumber={getNewWorkPackageNumber}
-            clearEdits={clearEdits}
-            setClearEdits={setClearEdits}
+            allWbsElements={allWbsElements}
+            showWorkPackagesMap={showWorkPackagesMap}
+            setShowWorkPackagesMap={setShowWorkPackagesMap}
+            removeAddedProjects={removeAddedProjects}
+            removeAddedWorkPackages={removeAddedWorkPackages}
           />
         ) : (
           <></>
