+++ src/frontend/src/pages/GanttPage/GanttChartTeamSection.tsx
@@ -2,49 +2,47 @@ import { Edit } from '@mui/icons-material';
 import { Box, Chip, IconButton, Typography, useTheme } from '@mui/material';
 import GanttChartSection from './GanttChartSection';
 import {
-  applyChangesToEvent,
+  aggregateGanttChanges,
+  applyChangesToWBSElement,
   GanttChange,
   RequestEventChange,
   transformProjectPreviewToProject
 } from '../../utils/gantt.utils';
-import { useEffect, useState } from 'react';
-import AddProjectModal from './GanttChartComponents/AddProjectModal';
+import { Dispatch, useState } from 'react';
 import useId from '@mui/material/utils/useId';
-import { Project, ProjectPreview, Team, WbsElementStatus, wbsPipe, WorkPackage } from 'shared';
+import { Project, ProjectPreview, Team, WbsElement, WbsElementStatus, wbsPipe, WorkPackage } from 'shared';
 import { projectWbsPipe } from '../../utils/pipes';
+import { GanttRequestChangeModal } from './GanttChartComponents/GanttChangeModals/GanttRequestChangeModal';
+import AddGanttProjectModal from './GanttChartComponents/AddGanttProjectModal';
 
 interface GanttChartTeamSectionProps {
   startDate: Date;
   endDate: Date;
-  saveChanges: (eventChanges: GanttChange[]) => void;
-  showWorkPackagesMap: Map<string, boolean>;
-  setShowWorkPackagesMap: React.Dispatch<React.SetStateAction<Map<string, boolean>>>;
   team: Team;
   projects: ProjectPreview[];
-  highlightedChange?: RequestEventChange;
+  addNewWorkPackage: (workPackage: WorkPackage) => void;
+  addNewProject: (project: Project) => void;
   getNewProjectNumber: (carNumber: number) => number;
-  addProject: (project: ProjectPreview) => void;
-  addWorkPackage: (workPackage: WorkPackage) => void;
-  getNewWorkPackageNumber: (projectId: string) => number;
-  clearEdits: boolean;
-  setClearEdits: (clearEdits: boolean) => void;
+  allWbsElements: WbsElement[];
+  showWorkPackagesMap: Map<string, boolean>;
+  setShowWorkPackagesMap: Dispatch<React.SetStateAction<Map<string, boolean>>>;
+  removeAddedProjects: (projects: Project[]) => void;
+  removeAddedWorkPackages: (workPackages: WorkPackage[]) => void;
 }
 
 const GanttChartTeamSection = ({
   startDate,
   endDate,
-  saveChanges,
-  showWorkPackagesMap,
-  setShowWorkPackagesMap,
   team,
   projects,
+  addNewWorkPackage,
+  addNewProject,
   getNewProjectNumber,
-  highlightedChange,
-  addProject,
-  addWorkPackage,
-  getNewWorkPackageNumber,
-  clearEdits,
-  setClearEdits
+  allWbsElements,
+  showWorkPackagesMap,
+  setShowWorkPackagesMap,
+  removeAddedProjects,
+  removeAddedWorkPackages
 }: GanttChartTeamSectionProps) => {
   const theme = useTheme();
   const [ganttChanges, setGanttChanges] = useState<GanttChange[]>([]);
@@ -52,14 +50,29 @@ const GanttChartTeamSection = ({
   const [showAddProjectModal, setShowAddProjectModal] = useState(false);
   const id = useId() || 'id';
   const [projectsState, setProjectsState] = useState<ProjectPreview[]>([...projects]);
+  const [requestEventChanges, setRequestEventChanges] = useState<RequestEventChange[]>([]);
+  const [addedProjects, setAddedProjects] = useState<Project[]>([]);
+  const [addedWorkPackages, setAddedWorkPackages] = useState<WorkPackage[]>([]);
 
-  useEffect(() => {
-    if (clearEdits) {
-      console.log('clearing edits', projects);
-      setProjectsState([...projects]);
-      setClearEdits(false);
-    }
-  }, [clearEdits, projects, setClearEdits]);
+  const teamSectionBackgroundStyle = {
+    mt: 1,
+    py: 1,
+    background: isEditMode ? theme.palette.divider : 'transparent',
+    borderRadius: '0.25rem',
+    width: 'fit-content'
+  };
+
+  const teamDescriptionContainerStyle = {
+    display: 'flex',
+    alignItems: 'center',
+    gap: 2,
+    mb: '-15px',
+    pl: 2,
+    position: 'sticky',
+    left: 0,
+    width: 'fit-content',
+    height: '30px'
+  };
 
   const createChange = (change: GanttChange) => {
     setGanttChanges([...ganttChanges, change]);
@@ -69,7 +82,12 @@ const GanttChartTeamSection = ({
     saveChanges(ganttChanges);
     setGanttChanges([]);
     setIsEditMode(false);
-    setClearEdits(false);
+  };
+
+  const handleCancel = () => {
+    setIsEditMode(false);
+    setGanttChanges([]);
+    setProjectsState([...projects]);
   };
 
   const handleEdit = () => {
@@ -85,86 +103,102 @@ const GanttChartTeamSection = ({
     project.workPackages.sort((a, b) => new Date(a.startDate).getTime() - new Date(b.startDate).getTime());
   });
 
-  const handleAddWorkPackage = (workPackage: WorkPackage) => {
+  const addNewWorkPackageHandler = (workPackage: WorkPackage) => {
     const project = projectsState.find((project) => wbsPipe(project.wbsNum) === projectWbsPipe(workPackage.wbsNum));
     if (!project) return;
     workPackage.wbsNum.workPackageNumber = project.workPackages.length + 1;
     project.workPackages = [...project.workPackages, workPackage];
     setProjectsState([...projectsState]);
-    addWorkPackage({ ...workPackage });
+    setAddedWorkPackages([...addedWorkPackages, workPackage]);
+    addNewWorkPackage(workPackage);
   };
 
-  const handleAddProject = (project: ProjectPreview) => {
+  const addNewProjectHandler = (project: ProjectPreview) => {
     setProjectsState([...projectsState, project]);
-    addProject(project);
+    const newProject: Project = transformProjectPreviewToProject(project, team);
+    setAddedProjects([...addedProjects, newProject]);
+    addNewProject(newProject);
   };
 
   const createChangeHandler = (change: GanttChange) => {
-    const parentProject = projectsState.find((project) => wbsPipe(project.wbsNum) === projectWbsPipe(change.element.wbsNum));
+    const parentProject = projectsState.find((project) => wbsPipe(project.wbsNum) === projectWbsPipe(change.element.wbsNum)); // Find the project that either the change is on, or the changes work package is a part of
     if (!parentProject) return;
 
-    const newProject: Project = applyChangesToEvent(
+    const newProject: Project = applyChangesToWBSElement(
       [change],
       transformProjectPreviewToProject(parentProject, team),
       parentProject
     ) as Project;
 
     const index = projectsState.findIndex((project) => wbsPipe(project.wbsNum) === projectWbsPipe(newProject.wbsNum));
-    projectsState[index] = newProject;
+    projectsState[index] = newProject; // Reset the project in the projects state, potentially unnecessary
     setProjectsState([...projectsState]);
     createChange(change);
   };
 
-  return (
-    <Box
-      sx={{
-        mt: 1,
-        py: 1,
-        background: isEditMode ? theme.palette.divider : 'transparent',
-        borderRadius: '0.25rem',
-        width: 'fit-content'
-      }}
-    >
-      <Box
-        sx={{
-          display: 'flex',
-          alignItems: 'center',
-          gap: 2,
-          mb: '-15px',
-          pl: 2,
-          position: 'sticky',
-          left: 0,
-          width: 'fit-content',
-          height: '30px'
+  const getNewWorkPackageNumber = (projectId: string) => {
+    const project = allWbsElements.find((wbsElement) => wbsElement.id === projectId) as Project;
+    if (!project) {
+      return 1;
+    }
+
+    return project.workPackages.length + 1;
+  };
+
+  const saveChanges = (eventChanges: GanttChange[]) => {
+    const requestEventChanges = aggregateGanttChanges(eventChanges, allWbsElements);
+
+    setRequestEventChanges(requestEventChanges);
+  };
+
+  const removeActiveModal = (changeId: string) => {
+    const newChanges = requestEventChanges.filter((change) => change.changeId !== changeId);
+    setRequestEventChanges(newChanges);
+    if (newChanges.length === 0) {
+      removeAddedProjects([...addedProjects]);
+      removeAddedWorkPackages([...addedWorkPackages]);
+      setAddedProjects([]);
+      setAddedWorkPackages([]);
+      setProjectsState([...projects]);
+    }
+  };
+
+  const AddProjectModal = () => {
+    return (
+      <AddGanttProjectModal
+        showModal={showAddProjectModal}
+        handleClose={() => setShowAddProjectModal(false)}
+        addProject={(project) => {
+          const newProject: ProjectPreview = {
+            id: id + projects.length + 1,
+            name: project.name,
+            wbsNum: {
+              carNumber: project.carNumber,
+              projectNumber: getNewProjectNumber(project.carNumber),
+              workPackageNumber: 0
+            },
+            status: WbsElementStatus.Inactive,
+            workPackages: []
+          };
+
+          addNewProjectHandler(newProject);
+
+          handleEdit();
+
+          createChange({
+            id,
+            type: 'create-project',
+            element: transformProjectPreviewToProject(newProject, team)
+          });
         }}
-      >
-        <AddProjectModal
-          showModal={showAddProjectModal}
-          handleClose={() => setShowAddProjectModal(false)}
-          addProject={(project) => {
-            const newProject: ProjectPreview = {
-              id: id + projects.length + 1,
-              name: project.name,
-              wbsNum: {
-                carNumber: project.carNumber,
-                projectNumber: getNewProjectNumber(project.carNumber),
-                workPackageNumber: 0
-              },
-              status: WbsElementStatus.Inactive,
-              workPackages: []
-            };
+      />
+    );
+  };
 
-            handleAddProject(newProject);
-
-            handleEdit();
-
-            createChange({
-              id,
-              type: 'create-project',
-              element: transformProjectPreviewToProject(newProject, team)
-            });
-          }}
-        />
+  return (
+    <Box sx={teamSectionBackgroundStyle}>
+      <AddProjectModal />
+      <Box sx={teamDescriptionContainerStyle}>
         <Typography variant="h6" fontWeight={400}>
           {team.teamName}
         </Typography>
@@ -172,15 +206,7 @@ const GanttChartTeamSection = ({
         {isEditMode ? (
           <Box display={'flex'} alignItems="center">
             <Chip label="Save" onClick={handleSave} sx={{ marginRight: '10px' }} />
-            <Chip
-              label="Cancel"
-              onClick={() => {
-                setIsEditMode(false);
-                setGanttChanges([]);
-                setProjectsState([...projects]);
-              }}
-              sx={{ marginRight: '10px' }}
-            />
+            <Chip label="Cancel" onClick={handleCancel} sx={{ marginRight: '10px' }} />
             <Chip label="Create Project" onClick={() => setShowAddProjectModal(true)} />
           </Box>
         ) : (
@@ -198,12 +224,15 @@ const GanttChartTeamSection = ({
           createChange={createChangeHandler}
           showWorkPackagesMap={showWorkPackagesMap}
           setShowWorkPackagesMap={setShowWorkPackagesMap}
-          highlightedChange={highlightedChange}
-          addWorkPackage={handleAddWorkPackage}
+          highlightedChange={requestEventChanges[requestEventChanges.length - 1]}
+          addWorkPackage={addNewWorkPackageHandler}
           teamName={team.teamName}
           getNewWorkPackageNumber={getNewWorkPackageNumber}
         />
       </Box>
+      {requestEventChanges.map((change) => (
+        <GanttRequestChangeModal change={change} open handleClose={() => removeActiveModal(change.changeId)} />
+      ))}
     </Box>
   );
 };
